test: Add a new test file that tests the whole program
The test checks if the file specified by the flag has been created. It then checks if the content of the file matches the expected result. To tun the test successfully you first have to run the main program using hello as your string input and output.txt as the value of your flag.